feature/cleanup: Unify and cleanup tasks
Migrated a number of tasks into their own py files.

Unified all of the tasks that require downloading artifacts
to have both a get_dependencies and install task.

Fix the debian add_repo_impl to use the Response.content
field of http responses and a BytesIO when downloading
repo keys which means that it will now work regardless
of whether the key is a binary or ASCII file.

Add package manager update before installing packages.

Refactor FEEDBACK so that it is a dict and not a list