Add support for debian12
Revamp how config overrides work.  Instead of a OVERRIDE or APPEND for
some types of configs it is a wholesale override of an entire config
block.  A base set of configs can be coupled with some overrides and the
result is all of the base configs plus the configs overridden and any
additional configs defined in the overrides.

The configs class now knows to look for an optional "base" key in the
distro config file.  If it is there, it will load the contents of the
base yaml file and merge the distro and base configs with the distro
configs taking precedence and overriding any values in the base distro
configs.

Added additional functions that are debian 12 specific.

Added requests retry Util along with file and directory exists utils.

Add ability to verify checksums on remote host.

Fixed virtualbox list extpack output parsing.

Moved most of the integration test function definitions to the base test
class as they are already abstracted enough to work with any distro.